# ugc/src/static/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: UGC Events API
  description: API для обработки пользовательских событий сервиса фильмов.
  version: 1.0.0
servers:
  - url: '/api/v1'
tags:
  - name: events
    description: "Пользовательские события"
  - name: grades
    description: "Оценка пользователей к фильму"
  - name: favorites
    description: "Закладки пользователей, избранные фильмы"
  - name: reviews
    description: "Рецензии пользователей к фильму"
paths:
  /events:
    post:
      tags:
        - events
      summary: Create a new user event
      parameters:
        - in: query
          name: key
          schema:
            type: string
            default: test_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'

      responses:
        '202':
          description: Successful operation
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden

  /grades:
    get:
      tags:
        - grades
      summary: Получение списка оценок к фильмам пользователя
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden

  /grades/{film_id}:
    post:
      tags:
        - grades
      summary: Создание новой оценки к фильму
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCreate'
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    get:
      tags:
        - grades
      summary: Получение оценки к фильму
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    patch:
      tags:
        - grades
      summary: Обновление оценки к фильму
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCreate'

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - grades
      summary: Удаление оценки к фильму
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

  /favorites:
    get:
      tags:
        - favorites
      summary: Получение списка фильмов в избранном пользователя
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Favorite'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

  /favorites/{film_id}:
    post:
      tags:
        - favorites
      summary: Добавление фильма в избранное пользователем
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    get:
      tags:
        - favorites
      summary: Получение фильма в избранном пользователя
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - favorites
      summary: Удаление фильма из избранного пользователем
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

  /reviews:
    get:
      tags:
        - reviews
      summary: Получение списка рецензий к фильмам пользователя
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

  /reviews/{film_id}:
    post:
      tags:
        - reviews
      summary: Создание рецензии на фильм
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewWrite'
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    get:
      tags:
        - reviews
      summary: Получение списка рецензий на фильм
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '400':
          description: Invalid request, missing required parameters
        '404':
          description: Not found
      security: []

    patch:
      tags:
        - reviews
      summary: Изменение рецензии на фильм
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewWrite'
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - reviews
      summary: Удаление рецензии на фильм
      parameters:
        - in: path
          name: film_id
          schema:
            type: string
            example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
            default: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid request, missing required parameters
        '403':
          description: Forbidden
        '404':
          description: Not found

components:
  securitySchemes:
    cookieAuth: # arbitrary name for the security scheme; will be used in the "security" key later
      type: apiKey
      in: cookie
      name: access_token
      description: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJiZjA1MTE5Ny0wODFjLTQyYWUtOWVmMy01NjZmNDlkNThhNzMiLCJyb2xlcyI6InN1YnNjcmliZXIiLCJleHAiOjE3NDg0MTQ0NzJ9.UmjE_eqTeqwz0KmlrIQw1HtjymuIQR4vpVNO2G5Z3hI"
  schemas:
    Event:
      required:
        - entity_type
        - entity
        - action
      type: object
      properties:
        service:
          type: string
          example: Movies
        timestamp:
          type: string
          format: date-time
          example: 2024-05-13 14:00:00
        entity_type:
          type: string
          example: Фильм
        entity:
          type: string
          example: Иван Васильевич меняет профессию
        action:
          type: string
          example: Начало просмотра

    Grade:
      type: object
      properties:
        id:
          type: integer
          example: 66624800d89be65ed7202397
        film_id:
          type: string
          example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
        rating:
          type: integer
          format: int32
          example: 10
        user_id:
          type: string
          example: bf051197-081c-42ae-9ef3-566f49d58a73
        timestamp:
          type: string
          example: Thu, 06 Jun 2024 23:36:32 GMT

    GradeCreate:
      required:
        - rating
      type: object
      properties:
        rating:
          type: integer
          format: int32
          example: 10

    Favorite:
      type: object
      properties:
        id:
          type: integer
          example: 66624800d89be65ed7202397
        film_id:
          type: string
          example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
        user_id:
          type: string
          example: bf051197-081c-42ae-9ef3-566f49d58a73
        timestamp:
          type: string
          example: Thu, 06 Jun 2024 23:36:32 GMT

    Review:
      type: object
      properties:
        id:
          type: string
          example: 666361e25b1b01d7066d8134
        film_id:
          type: string
          example: cf083ea9-aa9f-4dd0-b8ee-29393c1a0a31
        author:
          type: string
          example: bf051197-081c-42ae-9ef3-566f49d58a73
        text:
          type: string
          example: Рецензия к фильму.
        rating:
          type: integer
          example: 5
        timestamp:
          type: string
          example: Fri, 07 Jun 2024 16:58:50 GMT

    ReviewWrite:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          example: Рецензия к фильму.

security:
  - cookieAuth: []